[TASK] Change datetime field to bigint
To be able to handle datetime before 1901, the datetime field is changed to big int. (example imports of old articles)

+++ ext_tables.sql
@@ -31,7 +31,7 @@ CREATE TABLE tx_news_domain_model_news (
 	title tinytext,
 	teaser text,
 	bodytext mediumtext,
-	datetime int(11) DEFAULT '0' NOT NULL,
+	datetime bigint(20) DEFAULT '0' NOT NULL,
 	archive int(11) DEFAULT '0' NOT NULL,
 	author tinytext,
 	author_email tinytext,
